+++ src/index.ts
@@ -1,4 +1,2 @@
 export * from './built-data.js';
 export { type Config } from './config.js';
-export * from './modules.js';
-export * from './schema-loader.js';